Remove delegate request handler
So I had been partial about this but decided to remove it completely.

A delegate handler abstraction is an opinionated thing. Userland
can implement their own way of doing this sort of thing and at the
same time, a separate package for a delegate handler could be
provided.

Said package would implement the RequestHandler interface
meaning that it could be possible to use it with amphp-bootstrap
if for example it is a case of an AbstractDelegateHandler that
implements the interface and requires class extension for the
delegate method that it targets. In bootstrap, it would be passed
as the end-product class, which by inheritance implements the
RequestHandler interface.

It is mandatory to implement RequestHandler interface, either
in the form of a simple RequestHandler fed to a Router or a
Router implementation itself.

Therefore I see no reason to add delegate handler in amphp-bootstrap.

Signed-off-by: Theo Gotsopoulos <theogpl57@gmail.com>